Add social media links
I needed a better way to manage my social media links which may change
in the future. Usually one would write a massive if statement which
looks messy but by creating social-media-links.html in includes and a
file in data which defines all my links that can be called into a page.